% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_auxiliary_functions.R
\name{aux_string_to_coords}
\alias{aux_string_to_coords}
\title{Mechanical coordinate conversion}
\usage{
aux_string_to_coords(coord_string)
}
\arguments{
\item{coord_string}{character. A string containg potential geographic coordinates.}
}
\value{
list. Contain latitude and longitude as the first and second elements,
respectively.
}
\description{
Mechanically convert character strings containing geographic coordinates and convert to sets of numeric values. Useful as most LLM experience better and more consistent results when asked to return a single string instead of two separate values.
}
\details{
Will convert all strings to N, W for consistency's sake.
Future updates will probably make it a toggle.
}
\examples{
example = "19 ° 34 ’ S 29 ° 10 ° E"
aux_string_to_coords(example)
}
