% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_gazetteer.R
\name{gazetteer}
\alias{gazetteer}
\title{Get geographic coordinates from localities}
\usage{
gazetteer(locality)
}
\arguments{
\item{locality}{vector of characters. A vector of location names to process with a gazetteer.}
}
\value{
matrix.
}
\description{
Extract geographic coordinates from strings containing location names, using an online index.
through usage of a gazetteer for Python.
}
\details{
Gazetteers are geographical indexes, with the one used for \code{arete} being \href{https://nominatim.org/release-docs/develop/api/Overview/}{Nominatim} which is accessible through the \code{GeoPy} client.
}
\examples{
\dontrun{
example = c("Lisbon", "London")

gazzetteer(example)}
}
\seealso{
\code{\link{arete_setup}}
}
