% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_install_python_packages.R
\name{install_python_packages}
\alias{install_python_packages}
\title{Update python dependencies}
\usage{
install_python_packages(envname, verbose = TRUE, ...)
}
\arguments{
\item{envname}{character. Name or path to the Python virtual environment being used for arete.}

\item{verbose}{logical. Determines if output should be printed.}

\item{...}{Other parameters to pass to the installation function as per the documentation of \code{\link[reticulate]{py_install}}}
}
\value{
NULL. If any packages failed to install, a character listing which.
}
\description{
Install python dependencies needed to run arete.
}
\examples{
\dontrun{
install_python_packages("./path/to/venv")
}
}
\seealso{
\code{\link{arete_setup}}
}
