% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_labels_unique.R
\name{labels_unique}
\alias{labels_unique}
\title{Get the unique labels of a WebAnno document}
\usage{
labels_unique(data, info = "all")
}
\arguments{
\item{data}{character or WebAnnoTSV. The contents of a WebAnno TSV v3.3 file as
created by \code{\link{webanno_open}} or a path leading to it.}

\item{info}{character. A choice of "labels", "relations" and "all" on what to extract.}
}
\value{
list. With up to two elements, \code{$labels} and \code{$relations}, containing the respective elements.
}
\description{
Returns the unique classifications in the document
}
\examples{
example = arete_data("annotations")[[1]]
labels_unique(example, info = "all")
}
