% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_webanno_summary.R
\name{webanno_summary}
\alias{webanno_summary}
\title{Summarize the contents of a group of WebAnno tsv files}
\usage{
webanno_summary(input, goal = "relations", collected = TRUE, verbose = TRUE)
}
\arguments{
\item{input}{list of character or WebAnnoTSV. The contents of WebAnno TSV v3.3 files 
as created by \code{\link{webanno_open}} or a set of
paths leading to them.}

\item{goal}{character. One of "labels", "relations" or "all" determining what is summarized.}

\item{collected}{boolean. Organize the summary under a single dataframe, otherwise a list is returned.}

\item{verbose}{boolean. Print the final output.}
}
\value{
dataframe.
}
\description{
Returns
}
\examples{
example = arete_data("annotations")

webanno_summary(example)
}
