% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_webanno_creator.R
\docType{class}
\name{WebAnnoTSV-class}
\alias{WebAnnoTSV-class}
\alias{webanno_creator}
\title{WebAnno TSV v3.3 class creator.}
\description{
Crop raster layers to minimum size possible and uniformize NA values across layers.
}
